Issue #1664
In PR #1602 we decided to use unsigned short for storing RedisSock->port
but in previous release I reverted that change. In this PR I changed
signatire of redis_sock_create to prevent unneccecary convertion of types.

+++ library.c
@@ -1753,7 +1753,7 @@ PHP_REDIS_API void redis_debug_response(INTERNAL_FUNCTION_PARAMETERS, RedisSock 
  * redis_sock_create
  */
 PHP_REDIS_API RedisSock*
-redis_sock_create(char *host, int host_len, unsigned short port,
+redis_sock_create(char *host, int host_len, int port,
                   double timeout, double read_timeout,
                   int persistent, char *persistent_id,
                   long retry_interval)
